Visual effects and movement tweak

+++ Assets/Scripts/DrawYourHero/Damage/IDamageable.cs
@@ -4,6 +4,6 @@ namespace DrawYourHero.Damage
 {
     public interface IDamageable
     {
-        public void Damage(DamageSource source, GameObject sourceObject, float damage);
+        public void Damage(DamageSource source, GameObject sourceObject, int damage);
     }
 }
